statistical-transformation: modify stat-count to return new data-frame with only unique values.
The previous behavior was to return a copy of the incoming data-frame
and add a column to that with counts at each value of the :x aesthetic.
This produces a data-frame with (potentially) multiple :x values with
counts.

The new behaviour is to return a new data-frame with only unique values
of :x and the associated frequency of values at that location.